requirements.txt
huggingface_hub==0.30.2
PyYAML==6.0.2
shodan==1.31.0
torch==2.6.0
transformers==4.51.1
accelerate==1.6.0